backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.12
opencv-python-headless==4.10.0.84
Pillow==10.4.0
numpy==1.26.4
scipy==1.14.0
torch==2.4.0
torchvision==0.19.0
timm==1.0.9
transformers==4.44.0
pywt==1.7.0
